requirements.txt
pytest
pylint
pytest-cov
black
ipython
fastapi
supabase
uvicorn
pandas
pydantic
numpy
datetime
python-dotenv
transformers
pyyaml
requests
scikit-learn
yellowbrick
seaborn
matplotlib
plotly
langgraph
langchain
langchain_community
langchain_huggingface
langchain_google_genai
langchain_core
openai
sentry-sdk
statsmodels
pydantic[email]


#pytorch gpu version -> pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
